requirements.txt
couchbase==4.3.5
python-dotenv==1.1.0